refactor(AttendanceSystem): Controller 삭제
- 오직 자바 프로그램을 위해, 아키텍쳐의 의미 컨트롤러 레이어 삭제했습니다.
